BUG 2: More than normal memory usage is observed
Through code inspection it was noticed that over time we populate the sessions list
and we keep information about all sessions even the finished one.
Under pprof we noticed there  was a number of persistent and in-use allocated objects under
a certain allocation stack that kept increasing with the number of sessions that started and stopped
(even when gc was forced in pprof).
These were UUID string objects and this confirmed our initial assumption about the lingering session
information being the cause of the observed issue.
The fix was to remove the ended session's from the server's session list.
It was verified through pproof that the change led to those objects to not be kept around.